# backend/libs/svc_vector_contracts/src/svc_vector_contracts/models/__init__.py
# generated by datamodel-codegen:
#   filename:  openapi.patched.json
#   timestamp: 2026-01-18T16:31:23+00:00

from pydantic import UUID4, BaseModel, Field, RootModel


class QDrantVectorRequest(BaseModel):
    embedding: list[float] = Field(..., title="Embedding")


class QDrantVectorResponse(BaseModel):
    vector_ids: list[UUID4] = Field(..., title="Vector Ids")


class VectorMetaRequest(BaseModel):
    vector_ids: list[UUID4] = Field(..., title="Vector Ids")


class VectorProfileResponse(BaseModel):
    integration_id: UUID4 = Field(..., title="Integration Id")
    schema_name: str = Field(..., title="Schema Name")
    table_name: str = Field(..., title="Table Name")
    table_meta: str = Field(..., title="Table Meta")


class VectorStatsResponse(BaseModel):
    vector_count: int = Field(..., title="Vector Count")


class TableMeta(BaseModel):
    integration_id: UUID4 = Field(..., title="Integration Id")
    schema_name: str = Field(..., title="Schema Name")
    table_name: str = Field(..., title="Table Name")
    column_names: list[str] = Field(..., title="Column Names")


class ValidationError(BaseModel):
    loc: list[str | int] = Field(..., title="Location")
    msg: str = Field(..., title="Message")
    type: str = Field(..., title="Error Type")


class ResponseGetProfilesV1VectorOrganizationsOrganizationIdUsersUserIdIntegrationsIntegrationIdVectorsProfilesGet(
    RootModel[list[VectorProfileResponse]]
):
    root: list[VectorProfileResponse] = Field(
        ...,
        title="Response Get Profiles V1 Vector Organizations  Organization Id  Users  User Id  Integrations  Integration Id  Vectors Profiles Get",
    )


class HTTPValidationError(BaseModel):
    detail: list[ValidationError] | None = Field(default=None, title="Detail")


class VectorMetaResponse(BaseModel):
    integration_id: UUID4 = Field(..., title="Integration Id")
    schema_name: str = Field(..., title="Schema Name")
    table_name: str = Field(..., title="Table Name")
    table_meta: TableMeta


class ResponseGetMetaV1VectorOrganizationsOrganizationIdUsersUserIdIntegrationsVectorsMetaGet(
    RootModel[list[VectorMetaResponse]]
):
    root: list[VectorMetaResponse] = Field(
        ...,
        title="Response Get Meta V1 Vector Organizations  Organization Id  Users  User Id  Integrations Vectors Meta Get",
    )